feat: reorganize parallel optimizer
