+++ src/state/optimizer-parallel/modes/heuristics.ts
@@ -1,0 +1,72 @@
+import { Dispatch } from 'redux';
+import type { WorkerWrapper } from '../calculate';
+import { Combination, ResultData, Settings } from '../optimizerSetup';
+import { getLayerCombinations, getLayerNumber } from '../tree';
+import { getExtrasIdsCombinations, getTotalCombinations, splitCombinations } from '../utils';
+import {
+  START_HEURISTICS,
+  StartHeuristicsMessage,
+  isFinishHeuristicsMessage,
+} from '../worker/workerMessageTypes';
+import runCalcNormal from './normal';
+
+export default function runCalcHeuristics(
+  reduxState: any,
+  dispatch: Dispatch<any>,
+  workers: WorkerWrapper[],
+  settings: Settings,
+  maxThreads: number,
+) {
+  const extrasIds: string[][] = getExtrasIdsCombinations(reduxState);
+  // for a particular problem it might not be possible to split work efficiently into effectiveThreads chunks
+  const effectiveThreads = Math.min(maxThreads, getTotalCombinations(extrasIds, 1));
+
+  const layer = getLayerNumber(extrasIds, effectiveThreads);
+  const layerCombinations = getLayerCombinations(extrasIds, layer);
+  const chunks = splitCombinations(layerCombinations, effectiveThreads);
+
+  const combinations: Combination[] = [];
+  const resultData: ResultData[] = [];
+
+  workers.forEach((workerObj, index) => {
+    if (index >= effectiveThreads) {
+      return;
+    }
+
+    workerObj.worker.onmessage = (e) => {
+      if (isFinishHeuristicsMessage(e.data)) {
+        workerObj.status = 'finished_heuristics';
+        // check if all workers are finished
+        const allFinished = workers
+          .slice(0, effectiveThreads)
+          .every((locWorker) => locWorker.status === 'finished_heuristics');
+        const combs = e.data.data.map((c) => c[0]);
+        const res = e.data.data.map((c) => c[1]);
+        combinations.push(...combs);
+        resultData.push(...res);
+
+        if (allFinished) {
+          console.log("All workers finished heuristics, let's start the real calculation");
+          runCalcNormal(
+            reduxState,
+            dispatch,
+            workers,
+            combinations,
+            resultData,
+            settings,
+            maxThreads,
+            true,
+          );
+        }
+      }
+    };
+    const message: StartHeuristicsMessage = {
+      type: START_HEURISTICS,
+      chunks: chunks[index],
+      extrasIds,
+      reduxState,
+      settings,
+    };
+    workerObj.worker.postMessage(message);
+  });
+}
